Update README, split tests into multiple files, add benchmarks for ODESolver, add examples, fix typos. Set the version to 0.2.0-alpha
