+++ Cargo.toml
@@ -1,12 +1,12 @@
 [package]
 name = "eqsolver"
-version = "0.1.3"
+version = "0.2.0-alpha"
 edition = "2021"
 description = "A library that solves equations using numerical methods"
 repository = "https://github.com/AzeezDa/eqsolver"
 license = "MIT"
 readme = "README.md"
-keywords = ["equation", "solver", "mathematics", "ode"]
+keywords = ["equation", "solver", "mathematics", "ode", "optimization"]
 categories = ["science", "mathematics", "algorithms"]
 
 [badges]
@@ -21,7 +21,6 @@ nalgebra = "0.33.0"
 num-traits = "0.2.19"
 rand = "0.8.5"
 rand_distr = "0.4.3"
-
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports"] }
