+++ benchmarks/lib.rs
@@ -1,6 +1,7 @@
 use criterion::criterion_main;
 
 mod multi;
+mod ode;
 mod single;
 
-criterion_main!(single::benches, multi::benches);
+criterion_main!(single::benches, multi::benches, ode::benches);
