+++ src/tests/mod.rs
@@ -1,0 +1,4 @@
+pub mod single;
+pub mod multi;
+pub mod ode;
+pub mod misc;
